# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CreateRoomInput {
  /**
     * @minLength 1
     * @maxLength 18
     */
  playerName: string;
  /**
     * @minimum 3
     * @maximum 10
     */
  maxPlayers?: number;
}

export interface JoinRoomInput {
  /**
     * @minLength 1
     * @maxLength 18
     */
  playerName: string;
}

export interface RoomActionInput {
  /** @minLength 1 */
  playerId: string;
}

export interface PassCardInput {
  /** @minLength 1 */
  playerId: string;
  /** @minimum 0 */
  cardIndex: number;
}

export interface PlayerView {
  id: string;
  name: string;
  avatar: string;
  score: number;
  cardCount: number;
  isHost: boolean;
  isReady: boolean;
}

export type RoomViewStatus = typeof RoomViewStatus[keyof typeof RoomViewStatus];


export const RoomViewStatus = {
  waiting: 'waiting',
  playing: 'playing',
  finished: 'finished',
} as const;

export interface RoomView {
  roomCode: string;
  viewerPlayerId: string;
  status: RoomViewStatus;
  maxPlayers: number;
  round: number;
  /** @nullable */
  activePlayerId: string | null;
  /** @nullable */
  winnerId: string | null;
  players: PlayerView[];
  hand: string[];
  feed: string[];
  canStart: boolean;
  canPass: boolean;
  canCallZamrat: boolean;
}