# lib/api-zod/src/generated/types/createRoomInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRoomInput {
  /**
     * @minLength 1
     * @maxLength 18
     */
  playerName: string;
  /**
     * @minimum 3
     * @maximum 10
     */
  maxPlayers?: number;
}