# lib/api-zod/src/generated/types/joinRoomInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JoinRoomInput {
  /**
     * @minLength 1
     * @maxLength 18
     */
  playerName: string;
}