# lib/api-zod/src/generated/types/passCardInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PassCardInput {
  /** @minLength 1 */
  playerId: string;
  /** @minimum 0 */
  cardIndex: number;
}