# lib/api-zod/src/generated/types/playerView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerView {
  id: string;
  name: string;
  avatar: string;
  score: number;
  cardCount: number;
  isHost: boolean;
  isReady: boolean;
}