# lib/api-zod/src/generated/types/roomView.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerView } from './playerView';
import type { RoomViewStatus } from './roomViewStatus';

export interface RoomView {
  roomCode: string;
  viewerPlayerId: string;
  status: RoomViewStatus;
  maxPlayers: number;
  round: number;
  /** @nullable */
  activePlayerId: string | null;
  /** @nullable */
  winnerId: string | null;
  players: PlayerView[];
  hand: string[];
  feed: string[];
  canStart: boolean;
  canPass: boolean;
  canCallZamrat: boolean;
}